refactor(core): restructure global CSS structure and TypeScript interfaces
