--- src/routes/folks.ts
@@ -1,1311 +1,0 @@
-import { images } from "./imports";
-
-export type Field =
-    | "interactive"
-    | "designer"
-    | "maps"
-    | "mastermind"
-    | "hands";
-
-export interface FolkData {
-    name: string;
-    page: string;
-    type: Field | string;
-    image: string;
-}
-
-export interface Folk {
-    id: number;
-    data: FolkData;
-}
-
-export const folks: Folk[] = [
-    {
-        id: 1,
-        data: {
-            name: "Adam Pearce",
-            page: "https://roadtolarissa.com/",
-            type: "interactive",
-            image: images["/src/lib/images/adam-pearce.webp"].default,
-        },
-    },
-    {
-        id: 2,
-        data: {
-            name: "Adrián Blanco",
-            page: "https://adrianblancoramos.com/",
-            type: "maps",
-            image: images["/src/lib/images/adrian-blanco.webp"].default,
-        },
-    },
-    {
-        id: 3,
-        data: {
-            name: "Aileen Clarke",
-            page: "https://aileenclarke.github.io/",
-            type: "maps",
-            image: images["/src/lib/images/aileen-clarke.webp"].default,
-        },
-    },
-    {
-        id: 4,
-        data: {
-            name: "Alba Martín",
-            page: "https://albamart.in/",
-            type: "interactive",
-            image: images["/src/lib/images/alba-martin.webp"].default,
-        },
-    },
-    {
-        id: 5,
-        data: {
-            name: "Alberto Cairo",
-            page: "https://www.albertocairo.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/alberto-cairo.webp"].default,
-        },
-    },
-    {
-        id: 6,
-        data: {
-            name: "Alessia Musìo",
-            page: "https://alessiamusio.com/",
-            type: "designer",
-            image: images["/src/lib/images/alessia-musio.webp"].default,
-        },
-    },
-    {
-        id: 7,
-        data: {
-            name: "Alli Torban",
-            page: "https://www.allitorban.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/alli-torban.webp"].default,
-        },
-    },
-    {
-        id: 8,
-        data: {
-            name: "Alvin Chang",
-            page: "https://alvinschang.com/",
-            type: "interactive",
-            image: images["/src/lib/images/alvin-chang.webp"].default,
-        },
-    },
-    {
-        id: 9,
-        data: {
-            name: "Amelia Wattenberger",
-            page: "https://wattenberger.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/amelia-wattenberger.webp"].default,
-        },
-    },
-    {
-        id: 10,
-        data: {
-            name: "Amy Cesal",
-            page: "https://www.amycesal.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/amy-cesal.webp"].default,
-        },
-    },
-    {
-        id: 11,
-        data: {
-            name: "Andrea Giambelli",
-            page: "https://www.andreagiambelli.com/",
-            type: "interactive",
-            image: images["/src/lib/images/andrea-giambelli.webp"].default,
-        },
-    },
-    {
-        id: 12,
-        data: {
-            name: "Arjun Kakkar",
-            page: "https://arjunkakkkar.mmm.page/",
-            type: "interactive",
-            image: images["/src/lib/images/arjun-kakkar.webp"].default,
-        },
-    },
-    {
-        id: 13,
-        data: {
-            name: "Ashley Cai",
-            page: "https://www.ashleycai.com/",
-            type: "interactive",
-            image: images["/src/lib/images/ashley-cai.webp"].default,
-        },
-    },
-    {
-        id: 14,
-        data: {
-            name: "Ben Tanen",
-            page: "https://ben-tanen.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/ben-tanen.webp"].default,
-        },
-    },
-    {
-        id: 15,
-        data: {
-            name: "Brody Smith",
-            page: "https://brody.fyi/",
-            type: "maps",
-            image: images["/src/lib/images/brody-smith.webp"].default,
-        },
-    },
-    {
-        id: 17,
-        data: {
-            name: "Cam Rodriguez",
-            page: "https://camrodriguez.carrd.co/",
-            type: "maps",
-            image: images["/src/lib/images/cam-rodriguez.webp"].default,
-        },
-    },
-    {
-        id: 18,
-        data: {
-            name: "Cédric Sam",
-            page: "https://cedric.sam.name/",
-            type: "maps",
-            image: images["/src/lib/images/cedric-sam.webp"].default,
-        },
-    },
-    {
-        id: 19,
-        data: {
-            name: "Chris Canipe",
-            page: "https://chriscanipe.com/",
-            type: "interactive",
-            image: images["/src/lib/images/chris-canipe.webp"].default,
-        },
-    },
-    {
-        id: 20,
-        data: {
-            name: "Connor Rothschild",
-            page: "https://www.connorrothschild.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/connor-rothschild.webp"].default,
-        },
-    },
-    {
-        id: 21,
-        data: {
-            name: "Daniel Gonzalez Cappa",
-            page: "https://gcappa.com/",
-            type: "designer",
-            image: images["/src/lib/images/daniel-gonzalez-cappa.webp"].default,
-        },
-    },
-    {
-        id: 23,
-        data: {
-            name: "Diana Estefanía Rubio",
-            page: "https://www.dianaestefaniarubio.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/diana-estefania-rubio.webp"].default,
-        },
-    },
-    {
-        id: 24,
-        data: {
-            name: "Diana Yukari",
-            page: "https://dianayukari.com/",
-            type: "designer",
-            image: images["/src/lib/images/diana-yukari.webp"].default,
-        },
-    },
-    {
-        id: 25,
-        data: {
-            name: "Edit Gyenge",
-            page: "https://editgyenge.com/",
-            type: "designer",
-            image: images["/src/lib/images/edit-gyenge.webp"].default,
-        },
-    },
-    {
-        id: 26,
-        data: {
-            name: "Eleanor Lutz",
-            page: "https://eleanorlutz.com/",
-            type: "designer",
-            image: images["/src/lib/images/eleanor-lutz.webp"].default,
-        },
-    },
-    {
-        id: 27,
-        data: {
-            name: "Elena Mejía",
-            page: "https://elenaml.com/",
-            type: "interactive",
-            image: images["/src/lib/images/elena-mejia.webp"].default,
-        },
-    },
-    {
-        id: 29,
-        data: {
-            name: "Elliot Bentley",
-            page: "https://elliotbentley.com/",
-            type: "interactive",
-            image: images["/src/lib/images/elliot-bentley.webp"].default,
-        },
-    },
-    {
-        id: 30,
-        data: {
-            name: "Ethan Singer",
-            page: "https://ethansinger.me/",
-            type: "interactive",
-            image: images["/src/lib/images/ethan-singer.webp"].default,
-        },
-    },
-    {
-        id: 31,
-        data: {
-            name: "Evelina Parrou",
-            page: "https://www.behance.net/evelinaj",
-            type: "designer",
-            image: images["/src/lib/images/evelina-parrou.webp"].default,
-        },
-    },
-    {
-        id: 32,
-        data: {
-            name: "Federica Fragapane",
-            page: "https://www.behance.net/FedericaFragapane",
-            type: "mastermind",
-            image: images["/src/lib/images/federica-fragapane.webp"].default,
-        },
-    },
-    {
-        id: 33,
-        data: {
-            name: "Florent Lavergne",
-            page: "https://www.behance.net/lavergneflor",
-            type: "maps",
-            image: images["/src/lib/images/florent-lavergne.webp"].default,
-        },
-    },
-    {
-        id: 34,
-        data: {
-            name: "Gabrielle Merite",
-            page: "https://www.gabriellemerite.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/gabrielle-merite.webp"].default,
-        },
-    },
-    {
-        id: 35,
-        data: {
-            name: "Gemma Busquets",
-            page: "https://www.gemmabusquets.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/gemma-busquets.webp"].default,
-        },
-    },
-    {
-        id: 36,
-        data: {
-            name: "Giorgia Lupi",
-            page: "https://giorgialupi.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/giorgia-lupi.webp"].default,
-        },
-    },
-    {
-        id: 37,
-        data: {
-            name: "Gurman Bhatia",
-            page: "https://gurmanbhatia.com/",
-            type: "interactive",
-            image: images["/src/lib/images/gurman-bhatia.webp"].default,
-        },
-    },
-    {
-        id: 38,
-        data: {
-            name: "Han Huang",
-            page: "https://jashcny.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/han-huang.webp"].default,
-        },
-    },
-    {
-        id: 39,
-        data: {
-            name: "Ihar Yanouski",
-            page: "https://yanouski.com/",
-            type: "interactive",
-            image: images["/src/lib/images/ihar-yanouski.webp"].default,
-        },
-    },
-    {
-        id: 41,
-        data: {
-            name: "Irene de la Torre Arenas",
-            page: "https://irenedelatorrearenas.wordpress.com/",
-            type: "interactive",
-            image: images["/src/lib/images/irene-de-la-torre-arenas.webp"]
-                .default,
-        },
-    },
-    {
-        id: 42,
-        data: {
-            name: "Jacopo Poletto",
-            page: "https://jacopopoletto.com/",
-            type: "interactive",
-            image: images["/src/lib/images/jacopo-poletto.webp"].default,
-        },
-    },
-    {
-        id: 43,
-        data: {
-            name: "Jan Diehm",
-            page: "https://jandiehmdesigns.com/",
-            type: "interactive",
-            image: images["/src/lib/images/jan-diehm.webp"].default,
-        },
-    },
-    {
-        id: 44,
-        data: {
-            name: "Jared Wilber",
-            page: "https://www.jwilber.me/",
-            type: "interactive",
-            image: images["/src/lib/images/jared-wilber.webp"].default,
-        },
-    },
-    {
-        id: 45,
-        data: {
-            name: "Jeremy C.F. Lin",
-            page: "https://jeremycflin.github.io/",
-            type: "maps",
-            image: images["/src/lib/images/jeremy-lin.webp"].default,
-        },
-    },
-    {
-        id: 46,
-        data: {
-            name: "Jin Wu",
-            page: "https://jinnwuu.github.io/portfolio/",
-            type: "maps",
-            image: images["/src/lib/images/jin-wu.webp"].default,
-        },
-    },
-    {
-        id: 48,
-        data: {
-            name: "Joey Lee",
-            page: "https://jk-lee.com/",
-            type: "maps",
-            image: images["/src/lib/images/joey-lee.webp"].default,
-        },
-    },
-    {
-        id: 49,
-        data: {
-            name: "Jonathan Corum",
-            page: "https://13pt.com/corum/",
-            type: "maps",
-            image: images["/src/lib/images/jonathan-corum.webp"].default,
-        },
-    },
-    {
-        id: 50,
-        data: {
-            name: "Juan Pablo Bravo",
-            page: "https://www.linkedin.com/in/juanpablobravocl/",
-            type: "designer",
-            image: images["/src/lib/images/juan-pablo-bravo.webp"].default,
-        },
-    },
-    {
-        id: 51,
-        data: {
-            name: "Judit Bekker",
-            page: "https://www.juditbekker.com/",
-            type: "interactive",
-            image: images["/src/lib/images/judit-bekker.webp"].default,
-        },
-    },
-    {
-        id: 52,
-        data: {
-            name: "Kate Snow",
-            page: "https://katesnowdesign.com/",
-            type: "designer",
-            image: images["/src/lib/images/kate-snow.webp"].default,
-        },
-    },
-    {
-        id: 53,
-        data: {
-            name: "Katie Marriner",
-            page: "https://katiemarriner.com/",
-            type: "interactive",
-            image: images["/src/lib/images/katie-marriner.webp"].default,
-        },
-    },
-    {
-        id: 55,
-        data: {
-            name: "Kristin Baumann",
-            page: "https://www.kristin-baumann.com/",
-            type: "interactive",
-            image: images["/src/lib/images/kristin-baumann.webp"].default,
-        },
-    },
-    {
-        id: 56,
-        data: {
-            name: "Krisztina Szucs",
-            page: "https://krisztinaszucs.com/",
-            type: "interactive",
-            image: images["/src/lib/images/krisztina-szucs.webp"].default,
-        },
-    },
-    {
-        id: 57,
-        data: {
-            name: "Ksenia Orlova",
-            page: "https://www.behance.net/orlovakseniia",
-            type: "maps",
-            image: images["/src/lib/images/ksenia-orlova.webp"].default,
-        },
-    },
-    {
-        id: 58,
-        data: {
-            name: "Laura Castro",
-            page: "https://www.iamlauracastro.com",
-            type: "interactive",
-            image: images["/src/lib/images/laura-castro.webp"].default,
-        },
-    },
-    {
-        id: 59,
-        data: {
-            name: "Laura Martín",
-            page: "https://lmartingrx.github.io/porfolio/",
-            type: "interactive",
-            image: images["/src/lib/images/laura-martin.webp"].default,
-        },
-    },
-    {
-        id: 60,
-        data: {
-            name: "Laura Navarro Soler",
-            page: "https://www.linkedin.com/in/laura-navarro-soler/",
-            type: "interactive",
-            image: images["/src/lib/images/laura-navarro-soler.webp"].default,
-        },
-    },
-    {
-        id: 61,
-        data: {
-            name: "Leonardo Nicoletti",
-            page: "https://www.leonardonicoletti.com/",
-            type: "maps",
-            image: images["/src/lib/images/leonardo-nicoletti.webp"].default,
-        },
-    },
-    {
-        id: 62,
-        data: {
-            name: "Lindsey Poulter",
-            page: "https://www.lindseypoulter.com/",
-            type: "interactive",
-            image: images["/src/lib/images/lindsey-poulter.webp"].default,
-        },
-    },
-    {
-        id: 63,
-        data: {
-            name: "Lisa Charlotte Muth",
-            page: "https://lisacharlottemuth.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/lisa-charlotte-muth.webp"].default,
-        },
-    },
-    {
-        id: 64,
-        data: {
-            name: "Lisa Hornung",
-            page: "https://inside-numbers.com/",
-            type: "designer",
-            image: images["/src/lib/images/lisa-hornung.webp"].default,
-        },
-    },
-    {
-        id: 65,
-        data: {
-            name: "Marcelo Duhalde",
-            page: "https://www.behance.net/marcelodc",
-            type: "designer",
-            image: images["/src/lib/images/marcelo-duhalde.webp"].default,
-        },
-    },
-    {
-        id: 66,
-        data: {
-            name: "Marco Hernández",
-            page: "https://mhinfographics.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/marco-hernandez.webp"].default,
-        },
-    },
-    {
-        id: 67,
-        data: {
-            name: "Mariano Zafra",
-            page: "http://www.marianozafra.com/",
-            type: "maps",
-            image: images["/src/lib/images/mariano-zafra.webp"].default,
-        },
-    },
-    {
-        id: 68,
-        data: {
-            name: "Marie Patino",
-            page: "https://mariepatino.com/",
-            type: "interactive",
-            image: images["/src/lib/images/marie-patino.webp"].default,
-        },
-    },
-    {
-        id: 69,
-        data: {
-            name: "Marta Sierra",
-            page: "https://www.martaviz.nl/",
-            type: "hands",
-            image: images["/src/lib/images/marta-sierra.webp"].default,
-        },
-    },
-    {
-        id: 70,
-        data: {
-            name: "Martin Grandjean",
-            page: "https://www.martingrandjean.ch/",
-            type: "mastermind",
-            image: images["/src/lib/images/martin-grandjean.webp"].default,
-        },
-    },
-    {
-        id: 71,
-        data: {
-            name: "Martina Dossi",
-            page: "https://www.behance.net/martinadossi",
-            type: "designer",
-            image: images["/src/lib/images/martina-dossi.webp"].default,
-        },
-    },
-    {
-        id: 72,
-        data: {
-            name: "Martina Zunica",
-            page: "https://www.behance.net/martinazunica_/",
-            type: "mastermind",
-            image: images["/src/lib/images/martina-zunica.webp"].default,
-        },
-    },
-    {
-        id: 73,
-        data: {
-            name: "Matteo Palù",
-            page: "https://www.matteopalu.com/",
-            type: "designer",
-            image: images["/src/lib/images/matteo-palu.webp"].default,
-        },
-    },
-    {
-        id: 74,
-        data: {
-            name: "Matteo Riva",
-            page: "https://crockhaus.com/",
-            type: "designer",
-            image: images["/src/lib/images/matteo-riva.webp"].default,
-        },
-    },
-    {
-        id: 75,
-        data: {
-            name: "Matthew Daniels",
-            page: "https://matthewfdaniels.github.io/home/",
-            type: "interactive",
-            image: images["/src/lib/images/matthew-daniels.webp"].default,
-        },
-    },
-    {
-        id: 76,
-        data: {
-            name: "Matthias Stahl",
-            page: "https://www.higsch.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/matthias-stahl.webp"].default,
-        },
-    },
-    {
-        id: 77,
-        data: {
-            name: "Michela Lazzaroni",
-            page: "https://www.behance.net/michelaz/projects",
-            type: "designer",
-            image: images["/src/lib/images/michela-lazzaroni.webp"].default,
-        },
-    },
-    {
-        id: 78,
-        data: {
-            name: "Michelle McGhee",
-            page: "https://mmcghee.notion.site/Michelle-McGhee-b4e0346f98fd4cb988d025173c1e94a7/",
-            type: "interactive",
-            image: images["/src/lib/images/michelle-mcGhee.webp"].default,
-        },
-    },
-    {
-        id: 79,
-        data: {
-            name: "Mira Rojanasakul",
-            page: "https://mirarojanasakul.com/",
-            type: "maps",
-            image: images["/src/lib/images/mira-rojanasakul.webp"].default,
-        },
-    },
-    {
-        id: 80,
-        data: {
-            name: "Mohamad Waked",
-            page: "https://alhadaqa.com/",
-            type: "designer",
-            image: images["/src/lib/images/mohamad-waked.webp"].default,
-        },
-    },
-    {
-        id: 81,
-        data: {
-            name: "Nadieh Bremer",
-            page: "https://www.visualcinnamon.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/nadieh-bremer.webp"].default,
-        },
-    },
-    {
-        id: 82,
-        data: {
-            name: "Nadja Popovich",
-            page: "https://nadjapopovich.com/",
-            type: "maps",
-            image: images["/src/lib/images/nadja-popovich.webp"].default,
-        },
-    },
-    {
-        id: 83,
-        data: {
-            name: "Nathan Yau",
-            page: "https://flowingdata.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/nathan-yau.webp"].default,
-        },
-    },
-    {
-        id: 85,
-        data: {
-            name: "Nikita Rokotyan",
-            page: "https://rokotyan.com/",
-            type: "interactive",
-            image: images["/src/lib/images/nikita-rokotyan.webp"].default,
-        },
-    },
-    {
-        id: 86,
-        data: {
-            name: "Pablo Robles",
-            page: "http://www.josepablorobles.com/",
-            type: "maps",
-            image: images["/src/lib/images/pablo-robles.webp"].default,
-        },
-    },
-    {
-        id: 88,
-        data: {
-            name: "Peter Cook",
-            page: "https://www.peterrcook.com/",
-            type: "interactive",
-            image: images["/src/lib/images/peter-cook.webp"].default,
-        },
-    },
-    {
-        id: 90,
-        data: {
-            name: "Prinz Magtulis",
-            page: "https://www.prinzmagtulis.com/",
-            type: "interactive",
-            image: images["/src/lib/images/prinz-magtulis.webp"].default,
-        },
-    },
-    {
-        id: 92,
-        data: {
-            name: "Rachael Dottle",
-            page: "https://rachaeldottle.com/",
-            type: "interactive",
-            image: images["/src/lib/images/rachael-dottle.webp"].default,
-        },
-    },
-    {
-        id: 93,
-        data: {
-            name: "Rahul Mukherjee",
-            page: "https://www.rmukherjee.com/",
-            type: "interactive",
-            image: images["/src/lib/images/rahul-mukherjee.webp"].default,
-        },
-    },
-    {
-        id: 94,
-        data: {
-            name: "Rosie Ettenheim",
-            page: "https://www.rosieettenheim.com/",
-            type: "designer",
-            image: images["/src/lib/images/rosie-ettenheim.webp"].default,
-        },
-    },
-    {
-        id: 95,
-        data: {
-            name: "Ryan Best",
-            page: "https://ryanabest.com/",
-            type: "interactive",
-            image: images["/src/lib/images/ryan-best.webp"].default,
-        },
-    },
-    {
-        id: 98,
-        data: {
-            name: "Sarah Fawson",
-            page: "https://thedatasays.org/",
-            type: "designer",
-            image: images["/src/lib/images/sarah-fawson.webp"].default,
-        },
-    },
-    {
-        id: 99,
-        data: {
-            name: "Sarah Fossheim",
-            page: "https://fossheim.io/",
-            type: "mastermind",
-            image: images["/src/lib/images/sarah-fossheim.webp"].default,
-        },
-    },
-    {
-        id: 100,
-        data: {
-            name: "Sarah Kay Miller",
-            page: "https://sarahkaymiller.com/",
-            type: "designer",
-            image: images["/src/lib/images/sarah-kay-miller.webp"].default,
-        },
-    },
-    {
-        id: 101,
-        data: {
-            name: "Shirley Wu",
-            page: "https://shirleywu.studio/",
-            type: "interactive",
-            image: images["/src/lib/images/shirley-wu.webp"].default,
-        },
-    },
-    {
-        id: 104,
-        data: {
-            name: "Simon Scarr",
-            page: "https://www.simonscarr.com/",
-            type: "designer",
-            image: images["/src/lib/images/simon-scarr.webp"].default,
-        },
-    },
-    {
-        id: 105,
-        data: {
-            name: "Simone Betito",
-            page: "https://www.simonebetito.com/",
-            type: "interactive",
-            image: images["/src/lib/images/simone-betito.webp"].default,
-        },
-    },
-    {
-        id: 106,
-        data: {
-            name: "Simran Parwani",
-            page: "https://simranparwani.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/simran-parwani.webp"].default,
-        },
-    },
-    {
-        id: 107,
-        data: {
-            name: "Stefan Pullen",
-            page: "https://www.stefanpullen.com/",
-            type: "designer",
-            image: images["/src/lib/images/stefan-pullen.webp"].default,
-        },
-    },
-    {
-        id: 108,
-        data: {
-            name: "Stefania Guerra",
-            page: "https://www.behance.net/stegue88e47f",
-            type: "designer",
-            image: images["/src/lib/images/stefania-guerra.webp"].default,
-        },
-    },
-    {
-        id: 111,
-        data: {
-            name: "Taylor Maggiacomo",
-            page: "https://www.taylormaggiacomo.com/",
-            type: "designer",
-            image: images["/src/lib/images/taylor-maggiacomo.webp"].default,
-        },
-    },
-    {
-        id: 112,
-        data: {
-            name: "Thomas Oide",
-            page: "https://thomasoide.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/thomas-oide.webp"].default,
-        },
-    },
-    {
-        id: 113,
-        data: {
-            name: "Tiziana Alocci",
-            page: "https://www.tizianaalocci.com/",
-            type: "designer",
-            image: images["/src/lib/images/tiziana-alocci.webp"].default,
-        },
-    },
-    {
-        id: 116,
-        data: {
-            name: "Weiyi Cai",
-            page: "https://www.caiweiyi.com/",
-            type: "maps",
-            image: images["/src/lib/images/weiyi-cai.webp"].default,
-        },
-    },
-    {
-        id: 117,
-        data: {
-            name: "Wendy Shijia",
-            page: "https://wendyshijia.notion.site/",
-            type: "maps",
-            image: images["/src/lib/images/wendy-shijia.webp"].default,
-        },
-    },
-    {
-        id: 118,
-        data: {
-            name: "Will Chase",
-            page: "https://www.williamrchase.com/",
-            type: "interactive",
-            image: images["/src/lib/images/will-chase.webp"].default,
-        },
-    },
-    {
-        id: 119,
-        data: {
-            name: "Wilson Andrews",
-            page: "https://www.wilsonmarkandrews.com/",
-            type: "interactive",
-            image: images["/src/lib/images/wilson-andrews.webp"].default,
-        },
-    },
-    {
-        id: 120,
-        data: {
-            name: "Yaryna Serkez",
-            page: "https://yarynam.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/yaryna-serkez.webp"].default,
-        },
-    },
-    {
-        id: 121,
-        data: {
-            name: "Yi Zhe Ang",
-            page: "https://yizhe-ang.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/yi-zhe-ang.webp"].default,
-        },
-    },
-    {
-        id: 122,
-        data: {
-            name: "Zach Levitt",
-            page: "https://zachlevitt.com/",
-            type: "maps",
-            image: images["/src/lib/images/zach-levitt.webp"].default,
-        },
-    },
-    {
-        id: 123,
-        data: {
-            name: "Curran Kelleher",
-            page: "https://github.com/curran",
-            type: "mastermind",
-            image: images["/src/lib/images/curran-kelleher.webp"].default,
-        },
-    },
-    {
-        id: 124,
-        data: {
-            name: "Moritz Stefaner",
-            page: "https://truth-and-beauty.net/",
-            type: "interactive",
-            image: images["/src/lib/images/moritz-stefaner.webp"].default,
-        },
-    },
-    {
-        id: 125,
-        data: {
-            name: "Fernando Becerra",
-            page: "https://www.fernandobecerra.com/",
-            type: "interactive",
-            image: images["/src/lib/images/fernando-becerra.webp"].default,
-        },
-    },
-    {
-        id: 126,
-        data: {
-            name: "Harry Stevens",
-            page: "https://harryjstevens.com/",
-            type: "mastermind",
-            image: images["/src/lib/images/harry-stevens.webp"].default,
-        },
-    },
-    {
-        id: 127,
-        data: {
-            name: "Nami Sumida",
-            page: "https://namisumida.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/nami-sumida.webp"].default,
-        },
-    },
-    {
-        id: 128,
-        data: {
-            name: "Giulia Manganelli",
-            page: "https://www.giuliamanganelli.com/",
-            type: "designer",
-            image: images["/src/lib/images/giulia-manganelli.webp"].default,
-        },
-    },
-    {
-        id: 129,
-        data: {
-            name: "Shelly Cheng",
-            page: "https://shellyscheng.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/shelly-cheng.webp"].default,
-        },
-    },
-    {
-        id: 130,
-        data: {
-            name: "Vinicius Sueiro",
-            page: "https://vsueiro.com/",
-            type: "interactive",
-            image: images["/src/lib/images/vinicius-sueiro.webp"].default,
-        },
-    },
-    {
-        id: 131,
-        data: {
-            name: "Jorge Martinez Rey",
-            page: "https://datavizcowboy.com/",
-            type: "designer",
-            image: images["/src/lib/images/jorge-martinez-rey.webp"].default,
-        },
-    },
-    {
-        id: 132,
-        data: {
-            name: "Rodrigo Menegat",
-            page: "https://meneg.at/",
-            type: "interactive",
-            image: images["/src/lib/images/rodrigo-menegat.webp"].default,
-        },
-    },
-    {
-        id: 133,
-        data: {
-            name: "Mariana Villamizar",
-            page: "https://marianaviro.myportfolio.com/work",
-            type: "interactive",
-            image: images["/src/lib/images/mariana-villamizar.webp"].default,
-        },
-    },
-    {
-        id: 134,
-        data: {
-            name: "Ani Matevosian",
-            page: "https://www.animatevosian.com/",
-            type: "interactive",
-            image: images["/src/lib/images/ani-matevosian.webp"].default,
-        },
-    },
-    {
-        id: 135,
-        data: {
-            name: "Stephanie Wu",
-            page: "https://www.stephdesigns.net/",
-            type: "interactive",
-            image: images["/src/lib/images/stephanie-wu.webp"].default,
-        },
-    },
-    {
-        id: 136,
-        data: {
-            name: "Hesham Eissa",
-            page: "https://heshameissa.com/",
-            type: "interactive",
-            image: images["/src/lib/images/hesham-eissa.webp"].default,
-        },
-    },
-    {
-        id: 137,
-        data: {
-            name: "Amanda Shendruk",
-            page: "https://ashendruk.com/",
-            type: "designer",
-            image: images["/src/lib/images/amanda-shendruk.webp"].default,
-        },
-    },
-    {
-        id: 138,
-        data: {
-            name: "Monsicha Srisuantang",
-            page: "https://monsicha.info/",
-            type: "designer",
-            image: images["/src/lib/images/monsicha-srisuantang.webp"].default,
-        },
-    },
-    {
-        id: 139,
-        data: {
-            name: "HaoChe Hung",
-            page: "https://haochehung.com/",
-            type: "designer",
-            image: images["/src/lib/images/haoche-hung.webp"].default,
-        },
-    },
-    {
-        id: 140,
-        data: {
-            name: "Julian Hoffmann Anton",
-            page: "https://www.julian-hoffmann-anton.com/",
-            type: "maps",
-            image: images["/src/lib/images/julian-hoffmann-anton.webp"].default,
-        },
-    },
-    {
-        id: 141,
-        data: {
-            name: "Catalina Hildebrandt",
-            page: "https://www.catalinahildebrandt.com/",
-            type: "designer",
-            image: images["/src/lib/images/catalina-hildebrandt.webp"].default,
-        },
-    },
-    {
-        id: 142,
-        data: {
-            name: "Carl Churchill",
-            page: "https://churchillgeo.com/",
-            type: "maps",
-            image: images["/src/lib/images/carl-churchill.webp"].default,
-        },
-    },
-    {
-        id: 143,
-        data: {
-            name: "Thanh Nguyen",
-            page: "https://www.thanhxthu.com/",
-            type: "designer",
-            image: images["/src/lib/images/thanh-nguyen.webp"].default,
-        },
-    },
-    {
-        id: 144,
-        data: {
-            name: "Denise Lu",
-            page: "https://www.deniselu.com/",
-            type: "interactive",
-            image: images["/src/lib/images/denise-lu.webp"].default,
-        },
-    },
-    {
-        id: 145,
-        data: {
-            name: "Nina Krug",
-            page: "https://www.ninakrug.de/",
-            type: "designer",
-            image: images["/src/lib/images/nina-krug.webp"].default,
-        },
-    },
-    {
-        id: 146,
-        data: {
-            name: "Italo de Oliveira",
-            page: "https://italodoliva.com.br/",
-            type: "designer",
-            image: images["/src/lib/images/italo-de-oliveira.webp"].default,
-        },
-    },
-    {
-        id: 147,
-        data: {
-            name: "Sara I. Belled",
-            page: "https://saraibelled.github.io/porfolio/",
-            type: "interactive",
-            image: images["/src/lib/images/sara-i-belled.webp"].default,
-        },
-    },
-    {
-        id: 148,
-        data: {
-            name: "Harry Lewis-Irlam",
-            page: "https://www.heligraphica.com/",
-            type: "interactive",
-            image: images["/src/lib/images/harry-lewis-irlam.webp"].default,
-        },
-    },
-    {
-        id: 149,
-        data: {
-            name: "Youjin Shin",
-            page: "https://www.youjin.info/",
-            type: "interactive",
-            image: images["/src/lib/images/youjin-shin.webp"].default,
-        },
-    },
-    {
-        id: 150,
-        data: {
-            name: "Álvaro Valiño",
-            page: "https://www.alvarovalino.com/",
-            type: "designer",
-            image: images["/src/lib/images/alvaro-valino.webp"].default,
-        },
-    },
-    {
-        id: 151,
-        data: {
-            name: "Alyssa Fowers",
-            page: "https://www.alyssafowers.com/portfolio",
-            type: "designer",
-            image: images["/src/lib/images/alyssa-fowers.webp"].default,
-        },
-    },
-    {
-        id: 152,
-        data: {
-            name: "Jillian Vordick",
-            page: "https://www.jillianvordick.com/",
-            type: "interactive",
-            image: images["/src/lib/images/jillian-vordick.webp"].default,
-        },
-    },
-    {
-        id: 153,
-        data: {
-            name: "Attila Bátorfy",
-            page: "https://attilabatorfy.com/",
-            type: "maps",
-            image: images["/src/lib/images/attila-batorfy.webp"].default,
-        },
-    },
-    {
-        id: 154,
-        data: {
-            name: "Alberto Lucas López",
-            page: "https://lucasinfografia.com",
-            type: "designer",
-            image: images["/src/lib/images/alberto-lucas-lopez.webp"].default,
-        },
-    },
-    {
-        id: 155,
-        data: {
-            name: "Evan Applegate",
-            page: "https://evanapplegate.com/",
-            type: "maps",
-            image: images["/src/lib/images/evan-applegate.webp"].default,
-        },
-    },
-    {
-        id: 156,
-        data: {
-            name: "Chiqui Esteban",
-            page: "https://chiquiesteban.me/",
-            type: "interactive",
-            image: images["/src/lib/images/chiqui-esteban.webp"].default,
-        },
-    },
-    {
-        id: 157,
-        data: {
-            name: "Carolina Correa Bull",
-            page: "https://carocorrea.com/",
-            type: "designer",
-            image: images["/src/lib/images/carolina-correa-bull.webp"].default,
-        },
-    },
-    {
-        id: 158,
-        data: {
-            name: "Luis Melgar",
-            page: "https://www.lmelgar.me/",
-            type: "interactive",
-            image: images["/src/lib/images/luis-melgar.webp"].default,
-        },
-    },
-    {
-        id: 159,
-        data: {
-            name: "Szu Yu Chen",
-            page: "https://sabrinamochi.github.io/",
-            type: "interactive",
-            image: images["/src/lib/images/szu-yu-chen.webp"].default,
-        },
-    },
-    {
-        id: 160,
-        data: {
-            name: "Artur Galocha",
-            page: "https://arturgalocha.com/",
-            type: "interactive",
-            image: images["/src/lib/images/artur-galocha.webp"].default,
-        },
-    },
-].map((folk): Folk => {
-    const path = folk.data.image;
-    return {
-        ...folk,
-        data: {
-            ...folk.data,
-            image: images[path]?.default || path,
-        },
-    };
-});
